cleaning up casts to float.

+++ test/com/jexpect/ToBeFloatTest.java
@@ -8,36 +8,36 @@ public class ToBeFloatTest {
 
   @Test
   public void To_Be_Actual() throws Exception {
-    expect((float) 10).toBe(10.0f);
+    expect(10f).toBe(10.0f);
   }
 
   @Test(expected = IllegalArgumentException.class)
   public void To_Be_Actual_Fail() throws Exception {
-    expect((float) 12).toBe(9f);
+    expect(12f).toBe(9f);
   }
 
   @Test(expected = IllegalArgumentException.class)
   public void To_Be_Actual_Fail_When_Actual_Is_Null() throws Exception {
-    expect((float) 10).toBe(null);
+    expect(10f).toBe(null);
   }
 
   @Test
   public void To_Be_Less_Than() throws Exception {
-    expect((float) 1).toBeLessThan(9.0f);
+    expect(1f).toBeLessThan(9.0f);
   }
 
   @Test(expected = IllegalArgumentException.class)
   public void To_Be_Less_Than_Fail() throws Exception {
-    expect((float) 11).toBeLessThan(9.0f);
+    expect(11f).toBeLessThan(9.0f);
   }
 
   @Test
   public void To_Be_Greater_Than() throws Exception {
-    expect((float) 10).toBeGreaterThan(9.0f);
+    expect(10f).toBeGreaterThan(9.0f);
   }
 
   @Test(expected = IllegalArgumentException.class)
   public void To_Be_Greater_Than_Fail() throws Exception {
-    expect((float) 1).toBeGreaterThan(9.0f);
+    expect(1f).toBeGreaterThan(9.0f);
   }
 }
